python/requirements.txt
raven==4.2.3
click==2.1
python-dateutil==2.2
sh==1.09
Logbook==0.7.0
structlog==0.4.1
Flask==0.10.1
Flask-RESTful==0.2.12
flask-mongoengine==0.7.0
docopt==0.6.1
celery==3.1.14
redis==2.9.1
-e git+https://github.com/dotcloud/docker-py.git@master#egg=docker-py-0.2.3
